Made Whash more compatible with imagehash.py
The Wavelet hash implementation of imagehash.py downscales the image
with antialiasing, which caused discrepancies between hashes generated
by imagehash.py and this implemention.

Included the disintegration/imaging library which allows for providing a
resample filter in order to pick the Lanczos.

This has increased our matching with the imagehash.py implementation to
100% when tested against a sample of 700 images.